requirements.txt
python-dotenv~=0.14.0
discord.py~=1.5.1
asyncpg~=0.21.0
pandas~=1.1.2
pyot~=3.0.1
discord~=1.0.1
python-dateutil~=2.8.1